requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-storage-file-share
Flask
Flask-JSON
Flask-JWT-Extended
Flask-Session
python-dotenv
requests
pydantic
PyJWT
pyodbc
pytest
python-dotenv